requirements.txt
yfinance~=0.2.58
pandas~=2.2.3
matplotlib
pygame
backtrader~=1.9.78.123
alpaca-trade-api
pytz~=2025.2
numpy~=2.2.5
ta~=0.11.0
python-dotenv~=1.1.0
textual~=3.2.0
plotext~=5.3.2
requests~=2.32.3
rich~=14.0.0
playsound~=1.3.0